refactor: (프론트)fetch 함수 분리, 불필요 import 제거, useAuth 사용
- Feed와 PersonalDooingle 외 페이지들에 대한 작업
- fetch 함수 분리
- 불필요 import 제거
- useAuth 사용
