+++ dooingle-front/src/components/FollowingUser.jsx
@@ -1,25 +1,7 @@
 import {Link} from "react-router-dom";
 import SmallSubmitButton from "./button/SmallSubmitButton.jsx";
 import {useState} from "react";
-import axios from "axios";
-import {BACKEND_SERVER_ORIGIN} from "../env.js";
-
-async function fetchAddFollow(userLink) {
-  const response = await axios.post(
-    `${BACKEND_SERVER_ORIGIN}/api/follow/${userLink}`,
-    null,
-    {withCredentials: true},
-  );
-  return response.data;
-}
-
-async function fetchCancelFollow(userLink) { // TODO Feed에도 있는 함수, 추후 반드시 정리 필요
-  // TODO 팔로우 취소 후 목록 화면 어떻게 보여줄지 고민 (1) 목록에서 빼버린다. (2) 팔로우 취소 버튼 대신 팔로우 버튼을 보여준다.
-  const response = await axios.delete(`${BACKEND_SERVER_ORIGIN}/api/follow/${userLink}`, {
-    withCredentials: true,
-  });
-  return response.data;
-}
+import {fetchAddFollow, fetchCancelFollow} from "../fetch.js";
 
 export default function FollowingUser({userName, userLink, userProfileImageUrl, userDescription}) {
 
@@ -45,11 +27,11 @@ export default function FollowingUser({userName, userLink, userProfileImageUrl, 
           <Link to={`/personal-dooingles/${userLink}`} className="font-bold text-[#8692ff]">{userName}</Link>
         </div>
         <div className="flex items-center max-w-[50%]">
-          <span className="text-[#5f6368]">{userDescription}</span>
+          <span className="text-[0.875rem] text-[#5f6368]">{userDescription}</span>
         </div>
       </div>
-      {isFollowingUser && <SmallSubmitButton type="button" onClick={handleCancelFollowButton} className="max-h-[2rem]">팔로우 취소</SmallSubmitButton>}
-      {!isFollowingUser && <SmallSubmitButton type="button" onClick={handleAddFollowButton} className="max-h-[2rem]">팔로우</SmallSubmitButton>}
+      {isFollowingUser && <SmallSubmitButton type="button" onClick={handleCancelFollowButton} className="max-h-[2rem] font-normal">팔로우 취소</SmallSubmitButton>}
+      {!isFollowingUser && <SmallSubmitButton type="button" onClick={handleAddFollowButton} className="max-h-[2rem] font-normal">팔로우</SmallSubmitButton>}
     </div>
   );
 }
